requirements.txt
# Core
torch
transformers
datasets
tqdm

# Vision & image handling
Pillow
captcha

# Web API
Flask
flask-cors